ai_chatbot_project/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-cloud-bigquery
plotly
langchain
# langchain-openai # Removed
google-generativeai
chromadb
sentence-transformers
